Check RSpec's config.order when determining if a seed is used
If a test suite has `config.order = :random` in its RSpec config, then
the test ordering is random, but turbo_tests would not print a seed
value. In order to get the correct behavior, a redundant `--seed` value
needed to be given to `turbo_tests`.

The RSpec configuration of `config.order = :random` should now be
correctly respected, so that running `turbo_tests` without `--seed`
correctly prints the random seed used the same way that `rspec` does
without giving `--seed`.

Note that this specifically does not try to handle the case where a
custom ordering strategy has been given. Only the predefined strategy of
`:random` is checked.